+++ internal/command/android_test.go
@@ -11,6 +11,11 @@ func Test_makeAndroidContext(t *testing.T) {
 	vol, err := mockDefaultVolume()
 	require.Nil(t, err)
 
+	engine, err := MakeEngine(autodetectEngine)
+	if err != nil {
+		t.Skip("engine not found", err)
+	}
+
 	type args struct {
 		flags *androidFlags
 		args  []string
@@ -76,6 +81,8 @@ func Test_makeAndroidContext(t *testing.T) {
 					StripDebug:   true,
 					Package:      ".",
 					Keystore:     "/app/testdata/my.keystore",
+					Engine:       engine,
+					Env:          map[string]string{},
 				},
 			},
 			wantErr: false,
@@ -105,6 +112,8 @@ func Test_makeAndroidContext(t *testing.T) {
 					StripDebug:   true,
 					Package:      ".",
 					Keystore:     "/app/testdata/my.keystore",
+					Engine:       engine,
+					Env:          map[string]string{},
 				},
 			},
 			wantErr: false,