+++ internal/command/windows_test.go
@@ -11,6 +11,11 @@ func Test_makeWindowsContext(t *testing.T) {
 	vol, err := mockDefaultVolume()
 	require.Nil(t, err)
 
+	engine, err := MakeEngine(autodetectEngine)
+	if err != nil {
+		t.Skip("engine not found", err)
+	}
+
 	type args struct {
 		flags *windowsFlags
 		args  []string
@@ -41,7 +46,8 @@ func Test_makeWindowsContext(t *testing.T) {
 					ID:           "windows-amd64",
 					OS:           "windows",
 					Architecture: "amd64",
-					Env:          []string{"GOOS=windows", "GOARCH=amd64", "CC=x86_64-w64-mingw32-gcc"},
+					Engine:       engine,
+					Env:          map[string]string{"GOOS": "windows", "GOARCH": "amd64", "CC": "x86_64-w64-mingw32-gcc"},
 					LdFlags:      []string{"-H=windowsgui"},
 					DockerImage:  windowsImage,
 				},
@@ -68,7 +74,8 @@ func Test_makeWindowsContext(t *testing.T) {
 					ID:           "windows-386",
 					OS:           "windows",
 					Architecture: "386",
-					Env:          []string{"GOOS=windows", "GOARCH=386", "CC=i686-w64-mingw32-gcc"},
+					Engine:       engine,
+					Env:          map[string]string{"GOOS": "windows", "GOARCH": "386", "CC": "i686-w64-mingw32-gcc"},
 					DockerImage:  windowsImage,
 				},
 			},
@@ -94,7 +101,8 @@ func Test_makeWindowsContext(t *testing.T) {
 					ID:           "windows-amd64",
 					OS:           "windows",
 					Architecture: "amd64",
-					Env:          []string{"GOOS=windows", "GOARCH=amd64", "CC=x86_64-w64-mingw32-gcc"},
+					Engine:       engine,
+					Env:          map[string]string{"GOOS": "windows", "GOARCH": "amd64", "CC": "x86_64-w64-mingw32-gcc"},
 					LdFlags:      []string{"-X main.version=1.2.3", "-H=windowsgui"},
 					DockerImage:  windowsImage,
 				},
@@ -121,7 +129,8 @@ func Test_makeWindowsContext(t *testing.T) {
 					ID:           "windows-amd64",
 					OS:           "windows",
 					Architecture: "amd64",
-					Env:          []string{"GOOS=windows", "GOARCH=amd64", "CC=x86_64-w64-mingw32-gcc"},
+					Engine:       engine,
+					Env:          map[string]string{"GOOS": "windows", "GOARCH": "amd64", "CC": "x86_64-w64-mingw32-gcc"},
 					LdFlags:      []string{"-H=windowsgui"},
 					DockerImage:  "test",
 				},